+++ packages/expo-image/src/ExpoImage.web.tsx
@@ -95,6 +95,8 @@ export default function ExpoImage({
   recyclingKey,
   style,
   nativeViewRef,
+  accessibilityLabel,
+  tintColor,
   ...props
 }: ImageNativeProps) {
   const imagePlaceholderContentFit = placeholderContentFit || 'scale-down';
@@ -118,7 +120,6 @@ export default function ExpoImage({
         ({ onAnimationFinished }) =>
           (className, style) => (
             <ImageWrapper
-              {...props}
               ref={nativeViewRef as React.Ref<HTMLImageElement> | undefined}
               source={placeholder?.[0]}
               style={{
@@ -133,6 +134,10 @@ export default function ExpoImage({
               contentPosition={{ left: '50%', top: '50%' }}
               hashPlaceholderContentPosition={contentPosition}
               hashPlaceholderStyle={imageHashStyle}
+              accessibilityLabel={accessibilityLabel}
+              cachePolicy={cachePolicy}
+              priority={priority}
+              tintColor={tintColor}
             />
           ),
       ]
@@ -148,7 +153,6 @@ export default function ExpoImage({
     ({ onAnimationFinished, onReady, onMount, onError: onErrorInner }) =>
       (className, style) => (
         <ImageWrapper
-          {...props}
           ref={nativeViewRef as React.Ref<HTMLImageElement> | undefined}
           source={selectedSource || placeholder?.[0]}
           events={{
@@ -168,12 +172,17 @@ export default function ExpoImage({
           contentPosition={selectedSource ? contentPosition : { top: '50%', left: '50%' }}
           hashPlaceholderContentPosition={contentPosition}
           hashPlaceholderStyle={imageHashStyle}
-          accessibilityLabel={props.accessibilityLabel}
+          accessibilityLabel={accessibilityLabel}
+          tintColor={tintColor}
         />
       ),
   ];
   return (
-    <View ref={containerRef} dataSet={{ expoimage: true }} style={[{ overflow: 'hidden' }, style]}>
+    <View
+      ref={containerRef}
+      dataSet={{ expoimage: true }}
+      style={[{ overflow: 'hidden' }, style]}
+      {...props}>
       <AnimationManager transition={transition} recyclingKey={recyclingKey} initial={initialNode}>
         {currentNode}
       </AnimationManager>